flask_website/requirements.txt
asgiref==3.11.0
attrs==25.3.0
blinker==1.9.0
cattrs==25.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Django==5.2.9
docstring-to-markdown==0.17
exceptiongroup==1.3.0
Flask==3.1.2
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
jedi==0.19.2
jedi-language-server==0.45.1
Jinja2==3.1.6
lsprotocol==2023.0.1
MarkupSafe==3.0.3
microvenv==2025.0
packaging==25.0
parso==0.8.5
pygls==1.3.1
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tomli==2.3.0
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
zipp==3.23.0
